pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.77.0",
    "psycopg2-binary>=2.9.10",
    "flask-wtf>=1.2.2",
    "keras>=2.14.0",
    "scikit-learn>=1.6.1",
    "sqlalchemy>=2.0.40",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "kaggle>=1.7.4.2",
    "werkzeug>=3.1.3",
    "wtforms>=3.2.1",
    "numpy>=2.2.5",
    "tensorflow>=2.14.0",
    "opencv-python-headless==4.7.0.72",
    "opencv-python>=4.11.0.86",
    "trafilatura>=2.0.0",
    "requests>=2.32.3",
    "kagglehub>=0.3.12",
]
